refactor(models/tcm): inline ELIC-style codec wiring; move convert to examples
Mirror the stf.py restyle for TCM:

- TCM now subclasses SimpleVAECompressionModel (drops the hand-written
  forward/compress/decompress in favour of the inherited g_a ->
  latent_codec -> g_s path).
- The Family 1 entropy stack is wired inline in __init__: channel_context
  + per-slice latent_codec dict comprehensions feeding
  HyperpriorLatentCodec(... ChannelGroupsLatentCodec(...)). The per-slice
  SWAtten support transforms are built by a local swatten_factory closure
  passed to build_mean_scale_head. Drops _build_tcm_latent_codec +
  build_channel_slice_codec indirection.
- convert_upstream_tcm_state_dict + its helpers (_rename_msa_keys,
  _ensure_identity_attention_projection, _is_upstream_tcm_state_dict,
  _UPSTREAM_* constants) move to examples/convert_tcm_checkpoint.py;
  from_state_dict now only loads already-converted state dicts. The
  _infer_* architecture-inference helpers stay in tcm.py (from_state_dict
  uses them). State-dict paths unchanged — round-trip diff stays 0.

tests: TestTcm loads convert_upstream_tcm_state_dict from the example
script via the _load_convert_fn importlib helper added in the stf commit.
